Update csi driver and other components (#32)
* Update csi driver and other components

This updates the csi-driver to the latest version, and syncs the versions of the other components with the ones defined in the upstream azurefile-csi chart.

The `csi-attacher` component has been removed from the upstream chart, see https://github.com/kubernetes-sigs/azurefile-csi-driver/pull/1664 . Should it be removed from here as well? I couldn't find any info on why it was removed nor if it should be replaced by something else. For now I've bumped the version to the latest release.

As per https://github.com/giantswarm/giantswarm/issues/30774

* Sync chart manifests with upstream

This commit syncs all changes made in the upstream chart since the last update.

* Fix image pullPolicy value

* Update changelog